Backend/requirements.txt
Django==5.2.8
django-cors-headers==4.9.0
django-environ==0.12.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
gunicorn==23.0.0
pillow==12.0.0
psycopg2-binary==2.9.11